+++ lib/ivis_opengl/textdraw.h
@@ -51,8 +51,8 @@ public:
 	int width();
 	// Height (in points)
 	int height();
-	void render(Vector2i position, PIELIGHT colour, float rotation = 0.0f, int maxWidth = -1, int maxHeight = -1);
-	void render(int x, int y, PIELIGHT colour, float rotation = 0.0f, int maxWidth = -1, int maxHeight = -1) { render(Vector2i{x,y}, colour, rotation, maxWidth, maxHeight); }
+	void render(Vector2f position, PIELIGHT colour, float rotation = 0.0f, int maxWidth = -1, int maxHeight = -1);
+	void render(float x, float y, PIELIGHT colour, float rotation = 0.0f, int maxWidth = -1, int maxHeight = -1) { render(Vector2f{x,y}, colour, rotation, maxWidth, maxHeight); }
 	void renderOutlined(int x, int y, PIELIGHT colour, PIELIGHT outlineColour);
 	int aboveBase(); // (in points)
 	int belowBase(); // (in points)
@@ -129,7 +129,7 @@ void iV_TextShutdown();
 void iV_font(const char *fontName, const char *fontFace, const char *fontFaceBold);
 
 int iV_GetEllipsisWidth(iV_fonts fontID);
-void iV_DrawEllipsis(iV_fonts fontID, Vector2i position, PIELIGHT colour);
+void iV_DrawEllipsis(iV_fonts fontID, Vector2f position, PIELIGHT colour);
 
 int iV_GetTextAboveBase(iV_fonts fontID);
 int iV_GetTextBelowBase(iV_fonts fontID);
