+++ src/advvis.cpp
@@ -76,7 +76,7 @@ UDWORD	avGetObjLightLevel(BASE_OBJECT *psObj, UDWORD origLevel)
 {
 	float div = (float)psObj->visible[selectedPlayer] / 255.f;
 	unsigned int lowest = origLevel / START_DIVIDE;
-	unsigned int newLevel = div * origLevel;
+	unsigned int newLevel = static_cast<unsigned int>(div * origLevel);
 
 	if (newLevel < lowest)
 	{
